back/requirements.txt
flask==3.1.1
flask-cors==4.0.2
flask-sqlalchemy==3.0.5
flask-jwt-extended==4.5.3
flask-limiter==3.5.0
pyotp==2.9.0
qrcode[pil]==7.4.2
bcrypt==4.0.1
werkzeug==3.1.0
cryptography==44.0.1
psycopg2-binary==2.9.10
flask-mail==0.9.1
python-dotenv==1.1.1
bleach==6.2.0
flask_talisman==1.1.0
librouteros==3.4.1
requests==2.32.3
celery==5.5.3